+++ components/AzTable.vue
@@ -1,59 +1,57 @@
 <template>
+  {{ data }}
   <Table
     :loading="loading"
-    :columns="props.columns"
+    :columns="computedColumns"
     :data-source="data"
     :pagination="pagination"
     @change="handleTableChange"
     stripe
-    bordered
     auto-header-height
     prefixCls="az-table"
     ref="azTableRef"
   >
     <template #title>
-      <div class="flex-1" ></div>
-
+      <div class="flex-1"></div>
       <AzButton type="link" size="small" @click="emits('addToggle')" icon="tabler:plus">
         {{ t('add') }}
       </AzButton>
       <slot name="addForm" />
       <Divider type="vertical" />
-      <AzFullScreen @click="toggle" />
+      <AzFullScreen @click="toggleFullScreen" />
     </template>
 
     <template
-      #customFilterDropdown="{
-        setSelectedKeys,
-        selectedKeys,
-        confirm,
-        clearFilters,
-        column,
-      }"
+      #customFilterDropdown="{ setSelectedKeys, selectedKeys, confirm, clearFilters, column }"
     >
+      {{ state }}
       <div class="p-4">
-        <Input
-          ref="searchInput"
-          :placeholder="`Search ${column.dataIndex}`"
+        <RangePicker
+          v-if="isDateColumn(column.dataIndex)"
+          showToday
+          format="YYYY-MM-DD HH:mm"
+          class="w-80 mb-4"
+          :placeholder="[t('startDate'), t('endDate')]"
           :value="selectedKeys[0]"
-          style="width: 188px; margin-bottom: 8px; display: block"
-          @change="
-            (e) => setSelectedKeys(e.target.value ? [e.target.value] : [])
-          "
-          @pressEnter="handleSearch(selectedKeys, confirm, column.dataIndex)"
+          @change="(e) => setSelectedKeys(e ? [e] : [])"
+        />
+        <Input
+          v-else
+          ref="searchInput"
+          :placeholder="`جستجو در ${t(column.dataIndex)}`"
+          :value="selectedKeys[0]"
+          class="mb-4"
+          @change="(e) => setSelectedKeys(e.target.value ? [e.target.value] : [])"
         />
         <div class="grid grid-cols-2 gap-2">
           <AzButton
             type="primary"
-            size="small"
-            @click="handleSearch(selectedKeys, confirm, column.dataIndex)"
+            @click="() => handleSubmit(selectedKeys, confirm, column.dataIndex)"
             icon="tabler:search"
           >
-            {{ t("search") }}
+            {{ t('filter') }}
           </AzButton>
-          <AzButton size="small" @click="handleReset(clearFilters)"
-            >{{ t("reset") }}
-          </AzButton>
+          <AzButton @click="() => handleReset(clearFilters)">{{ t('reset') }}</AzButton>
         </div>
       </div>
     </template>
@@ -66,85 +64,47 @@
       <template v-if="column.key === 'enabled'">
         <AzStatus :status="record.enabled" />
       </template>
-      <slot name="amir" :column="column" :record="record" />
+      <slot name="bodyCellSlot" :column="column" :record="record" />
     </template>
   </Table>
 </template>
+
 <script setup lang="ts">
-import type {
-  ColumnsType,
-  TablePaginationConfig,
-  TableProps,
-} from "ant-design-vue/es/table";
+import type { ColumnsType, TablePaginationConfig, TableProps } from 'ant-design-vue/es/table'
 import type {
   FilterValue,
   SorterResult,
-  TableCurrentDataSource,
-} from "ant-design-vue/lib/table/interface";
-import { Table, Divider, Input } from "ant-design-vue/es";
-import { useAxios } from "@/core/services/axios";
-import { useFullscreen, useOffsetPagination } from "@vueuse/core";
-import { ref, onMounted, reactive } from "vue";
-import { formatDate } from "@/core/utils";
-import { AzButton, AzFullScreen, AzStatus } from "@/core/components";
-import { Icon } from "@iconify/vue/dist/iconify.js";
-import { useI18n } from "vue-i18n";
+  TableCurrentDataSource
+} from 'ant-design-vue/lib/table/interface'
+import { Table, Divider, Input, RangePicker } from 'ant-design-vue/es'
+import { useAxios } from '@/core/services/axios'
+import { useFullscreen } from '@vueuse/core'
+import { ref, onMounted, reactive, computed } from 'vue'
+import { formatDate } from '@/core/utils'
+import { AzButton, AzFullScreen, AzStatus } from '@/core/components'
+import { Icon } from '@iconify/vue/dist/iconify.js'
+import { useI18n } from 'vue-i18n'
+import { isString, truncate } from 'lodash'
+import type { Dayjs } from 'dayjs'
+import dayjs from 'dayjs'
 
 interface Props {
-  url: string;
-  columns: ColumnsType;
+  url: string
+  columns: ColumnsType
 }
-const azTableRef = ref<HTMLElement | null>(null);
 
-const { toggle } = useFullscreen(azTableRef);
+const azTableRef = ref<HTMLElement | null>(null)
+const { toggle: toggleFullScreen } = useFullscreen(azTableRef)
+const props = defineProps<Props>()
+const emits = defineEmits(['addToggle'])
+const axios = useAxios()
+const data = ref([])
+const loading = ref(false)
+const { t } = useI18n()
 
-const props = withDefaults(defineProps<Props>(), {
-  url: "",
-  columns: () => [],
-});
- const emits =defineEmits(['addToggle'])
-const axios = useAxios();
-const data = ref([]);
-const loading = ref(false);
-const { t } = useI18n();
-const fetchData = async () => {
-  loading.value = true;
-
-  try {
-    const result = await axios.post({
-      url: props.url,
-      data: {
-        size: pagination.pageSize,
-        page: pagination.current - 1,
-        sortDir: sort.dir,
-        sortBy: sort.by,
-        [state.searchedColumn]: state.searchText,
-      },
-    });
-    data.value = result.content.map((item: any, index: number) => {
-      if (item["createdAt"]) {
-        item["createdAt"] = formatDate(item["createdAt"]);
-      }
-      if (item["updatedAt"]) {
-        item["updatedAt"] = formatDate(item["updatedAt"]);
-      }
-      if (item["expiredAt"]) {
-        item["expiredAt"] = formatDate(item["expiredAt"]);
-      }
-
-      return {
-        ...item,
-        row: index + 1,
-        key: item.id,
-      };
-    });
-    pagination.total = result.totalElements;
-  } catch (error) {
-    console.error("Error fetching data:", error);
-  } finally {
-    loading.value = false;
-  }
-};
+const computedColumns = computed(() => {
+  return [{ title: t('row'), dataIndex: 'row', key: 'row' }, ...props.columns]
+})
 
 const pagination = reactive({
   showSizeChanger: true,
@@ -153,65 +113,123 @@ const pagination = reactive({
   pageSizeOptions: [5, 10, 20, 30, 50],
   current: 1,
   pageSize: 5,
-  total: 0,
-});
+  total: 0
+})
+// const offsetPagination = useOffsetPagination({
+//   total: pagination.total,
+//   pageSize: pagination.defaultPageSize,
+//   onPageChange: (newPage) => {
+//     pagination.current = newPage.currentPage
+//     fetchData()
+//   },
+//   onPageSizeChange: (newSize) => {
+//     pagination.pageSize = newSize.currentPageSize
+//     pagination.current = 1
+//     fetchData()
+//   }
+// })
 
-const { currentPage, currentPageSize } = useOffsetPagination({
-  onPageChange: (_data) => {
-    pagination.current = _data.currentPage;
-    pagination.pageSize = _data.currentPageSize;
-  },
-  onPageSizeChange: (_data) => {
-    pagination.current = _data.currentPage;
-    pagination.pageSize = _data.currentPageSize;
-  },
-});
-
-const sort = reactive({
+const sort = reactive<any>({
   dir: undefined,
-  by: undefined,
-});
+  by: undefined
+})
 
-const handleTableChange: TableProps["onChange"] = (
+const state = reactive<any>({
+  value: undefined,
+  label: undefined
+})
+
+const fetchData = async () => {
+  loading.value = true
+  try {
+    const { content, totalElements } = await axios.post({
+      url: props.url,
+      data: {
+        size: pagination.pageSize,
+        page: pagination.current - 1,
+        sortDir: sort.dir,
+        sortBy: sort.by,
+        [state.label]: state.value
+      }
+    })
+
+    data.value = content.map((item: any, index: number) => ({
+      ...item,
+      ...Object.keys(item).reduce((acc: any, key) => {
+        if (isString(item[key])) acc[key] = truncate(item[key], { length: 30 })
+        return acc
+      }, {}),
+      row: (pagination.current - 1) * pagination.pageSize + index + 1,
+      key: item.id,
+      createdAt: item.createdAt ? formatDate(item.createdAt) : undefined,
+      updatedAt: item.updatedAt ? formatDate(item.updatedAt) : undefined,
+      expiredAt: item.expiredAt ? formatDate(item.expiredAt) : undefined
+    }))
+
+    pagination.total = totalElements
+  } catch (error) {
+    console.error('Error fetching data:', error)
+  } finally {
+    loading.value = false
+  }
+}
+
+const handleTableChange: TableProps['onChange'] = (
   pag: TablePaginationConfig,
   filters: Record<string, FilterValue | null>,
   sorter: SorterResult<any> | SorterResult<any>[],
   extra: TableCurrentDataSource<any>
 ) => {
-  pagination.current = pag.current ?? 1;
-  pagination.pageSize = pag.pageSize ?? 1;
-  sort.dir = sorter.order === "ascend" ? "asc" : "desc";
-  sort.by = sorter.columnKey;
-  fetchData();
-};
+  pagination.current = pag.current ?? 1
+  pagination.pageSize = pag.pageSize ?? 5
+  sort.dir = Array.isArray(sorter) ? sorter[0]?.order : sorter?.order
+  sort.dir = sort.dir === 'ascend' ? 'asc' : 'desc'
+  if (sort.dir) {
+    sort.by = Array.isArray(sorter) ? sorter[0]?.columnKey : sorter?.columnKey
+  } else {
+    sort.by = undefined
+  }
+  fetchData()
+}
 
-const state = reactive<any>({
-  searchText: undefined,
-  searchedColumn: undefined,
-});
-const handleSearch = (
-  selectedKeys: string[],
-  confirm: () => void,
-  dataIndex: string
-) => {
-  confirm();
-  state.searchText = selectedKeys[0];
-  state.searchedColumn = dataIndex;
-  fetchData();
-};
+const resetState = () => {
+  state.value = undefined
+  state.label = undefined
+}
+const handleTimeSubmit = (selectedKeys: string[], confirm: () => void, dataIndex: string) => {
+  console.log('🚀 ~ handleTimeSubmit ~ selectedKeys:', selectedKeys)
+  confirm()
+  state.value = selectedKeys[0]
+    ? [dayjs(selectedKeys[0][0]).format(), dayjs(selectedKeys[0][1]).format()]
+    : undefined
+  state.label = dataIndex
+  fetchData()
+}
+
+const handleSubmit = (selectedKeys: string[], confirm: () => void, dataIndex: string) => {
+  console.log("🚀 ~ handleSubmit ~ selectedKeys:", selectedKeys)
+  confirm()
+  state.value = selectedKeys[0]
+  state.label = dataIndex
+  fetchData()
+}
 
 const handleReset = (clearFilters: (arg0: { confirm: boolean }) => void) => {
-  clearFilters({ confirm: true });
-  state.searchText = undefined;
-  state.searchedColumn = undefined;
-  fetchData();
-};
-onMounted(async () => await fetchData());
+  clearFilters({ confirm: true })
+  resetState()
+  fetchData()
+}
+
+const isDateColumn = (dataIndex: string) => ['createdAt', 'updatedAt'].includes(dataIndex)
+
+onMounted(fetchData)
 </script>
+
 <style lang="less">
 .az-table {
   .az-table-title {
-    @apply flex  items-center;
+    display: flex;
+    align-items: center;
   }
 }
 </style>