+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
-
+## [0.0.4] - 2024-05-26
+### Added
+- table search by date
+### Fixed
+- table performance
 
 ## [0.0.3] - 2024-05-25
 ### Added